refactor(db): deliver write-path KMS via DB-held handle, not thread-local (race-free)
